# [warn: perf regression] Use `Bytes`