Conclusion:  Using `bytes` for hot paths seems to slow down quite measurably...


## Benchmark Comparison: Bytes vs &[u8] / Vec<u8>

### writers_vs_lock (criterion)

| Test           | With Bytes        | With &[u8] / Vec<u8> | Change              |
|----------------|------------------|----------------------|---------------------|
| 128 bytes      | 110–113 ms       | 108–110 ms           | −0.16% to −2.21%    |
| 4096 bytes     | 221–227 ms       | 217–221 ms           | −0.56% to −3.83%    |
| 65536 bytes    | 779–790 ms       | 775–784 ms           | −0.53% to −1.42%    |

### storage_benchmark.rs (custom)

| Operation         | With Bytes       | With &[u8] / Vec<u8> | Change           |
|-------------------|------------------|----------------------|------------------|
| Writes            | 2,042,067.8/s    | 3,375,536.2/s        | +65.3%           |
| Sequential reads  | 10,109,521.4/s   | 10,125,548.8/s       | ~0%              |
| Random reads      | 2,217,144.9/s    | 2,430,574.9/s        | +9.6%            |
| Batch reads       | 2,567,538.4/s    | 4,808,404.5/s        | +87.3%           |